Proudly presenting excellent test coverage for HAML.
C0 is at 99.0%... all except for one experimental feature.

So, now any changes have something to be tested against which should be a fairly strong background.

Now, the only problem left is that whitespace active areas get fucked. That's not a good thing. Once we get a 
philosophical stance on such subjects, then we shall figure crap out about what that all means.



git-svn-id: svn://hamptoncatlin.com/haml/trunk@13 7063305b-7217-0410-af8c-cdc13e5119b9