Split up the tests into two to match the new lib files that were created a few days ago.
Also, created tests for the propper "flattening" of whitespace-active-areas and created most of the code.

However, it is to be noted in this release... TESTS ARE BROKEN.

The flattening needs some love still. And the testing needs to get expanded further. I thought this would be easy, 
turns out its a bit of a bitch.



git-svn-id: svn://hamptoncatlin.com/haml/trunk@16 7063305b-7217-0410-af8c-cdc13e5119b9