Added an :output option to Haml, to determine which HTML variant it outputs.
By default, this is :xhtml. While :html could be specified, as is
written currently, will generate HTML 4.01 for anthing other than
:xhtml. Essentially, the only consequences of this change, are atomic
tags are not self closed for :html (ex. <br> instead of <br />), and
trying to issue a !!! XML will raise an error if you've specified an
output other than :xhtml.