Add a custom indent-region function for haml-mode.
The most common use of indent-region is to re-indent a region of code
that's been moved from some other document/location. Since Haml
indentation isn't deterministic, indenting each line doesn't work all
the time, so instead we assume the the relative indentation of the
region is correct. This means that only the first line needs to be
indented using haml-compute-indentation, and the rest of the region
can be indented with the same relative indentation.