Revert "Don't render unsightly XHTML newline escapes."
This reverts commit 0d4010cbaaf81541a2088a07acbbb8f723c55775.

This whole idea falls down when the toplevel template isn't necessarily active
when nested templates are being rendered.
Unfortunately, this is exactly the case for Rails and Merb,
which render the page template and then give the content to the layout.

This revert preserves some of the documentation added in the reverted commit.