Fix cache-marshalling errors on Windows.
On Windows, if a file is opened as text and then passed to Marshal.load
the call will fail if input has been read, per this thread:
http://blade.nagaokaut.ac.jp/cgi-bin/scat.rb/ruby/ruby-talk/96123 .
This change opens the cache files in binary mode.

After that change marshalling wouldn't load the file correctly.
I fixed this problem by opening the file for binary when writing the cache file.