[Sass] Check for dependencies using the parse tree for the Sass file
This means dependency checking will work for SCSS as well.
This may take more time than the quick-and-dirty regexp method,
but I don't think there's a good way around it.
Hopefully this will hit the cache most of the time.
We could speed it up a little by storing the dependencies directly in the cache,
or even more by keeping the dependencies in-memory.