[Haml] Optimize capture_haml for ugly mode.
There's no need to trim leading spaces in ugly mode.  We don't care what
the tab level is, and it's always going to be zero anyway in ugly mode.

In my own tests, I'm seeing a 33% performance improvement in rendering
certain pages with this change.  Also, in general I've found
capture_haml to perform poorly in profiling tests.