Allow Sass Imports to be looked up using arbitrary backends. These backends must adhere to the interface specified by Sass::Importer.
A Sass Importer is passed a name to import and a stylesheet context.
The importer must return a SassFile object if it finds it.

Sass Importers may be placed onto the :load_paths array. They will be
queried in order for imports with the exception being that the
importer that imported the current file is queried first if it can be.

These changes are completely backwards compatible.