[Sass] Make sure :filename isn't relative.
When we construct engines in importers, we need to be sure that the
value we pass in for the :filename option is relative to the importer
root. Otherwise, we won't know what it's relative to when we look at
it later.

I think this should make absolute paths work, too.