Modify the expected results for test_nested_media_around_properties.
Even if the @media appears between properties, it's okay to render all
the properties before the @media in the CSS. At least, that's how
nested rules have always worked, and if we change one we should change
both. It's an obscure enough case that I don't anticipate it mattering
much, though.