Change the edge-gem versioning scheme.
Rather than incrementing an index each time we push, we now impose an
indexing based on the repo structure. The version of the prerelease
gem is defined to be the number of commits since the stable
branch (precisely, the number of commits in `git log stable..master
--first-parent`). This means that when someone `rake install`s Sass
from a Git repo, the gem is installed with the proper version number.