[Sass] Fix a performance bug.
Backtracking in the SCSS parser when deciding between a declaration or a ruleset
was taking way too long. When an expected token wasn't found, the parser
constructed the full exception, which involved parsing the template for the
error message. This parsing was O(template size), leading to nesting-heavy
templates approaching O((template size)^2).

The new approach makes backtracking more light-weight, at the cost of making
exception-raising even more expensive (a perfectly acceptable tradeoff). When
backtracking is being considered, no exception is raised; instead, Kernel#throw
is used. Since the full error object is unavailable, once the parser knows which
error it wants to raise, it actually moves backwards and re-runs the production
that caused the error in order to re-raise the error.

Closes gh-327