Glob args for functions
In CSS2 functions have simple signatures: fixed number of arguments or optional
trailing arguments. This was implemented in SASS for a long time now. But CSS3
intruduces some new features that require variable arguments number. For
example, linear-gradient function. It expects an optional point or angle
arguments and a set of color stops. The number of color stops is limited only on
the lower end.

The need for variable argument number is expressed in Compass
framework. For example, take a lok at gradients implementation in v0.11.
Currently mixins have signaures of 10 arguments for color stops. And you're in
trouble if you want to have more.

This commit implements variable arguments number only for functions. It's more a
proof of concept. I plan to implement it for mixins too.

WIP