Add `adjust-by` function to manipulate units and numbers.
This function allows any number, incl. those with units, to be
manipulated by adding to/subtracting from them. Example:

    adjust-by(10px, 10%) => 11px

In template design, it is very often the case a unit is selected
and all computations then cast between the selected unit. A very
hacky approach already exists:

    $base: 10px;
    ($base / 1px) * 1.1 * 1px => 11px

but is less than ideal.