Change mixin name from "iphone" to "smartphone"
While mixin content blocks is an awesome feature, detecting specific devices/OS:s in CSS is error prone and a practice that should not be encouraged.

The example in the change log suggests detecting the iPhone by checking the viewport width, which is an incorrect approach. Even if examples are not meant to be used as is in production, it is still likely that they will influence developers.

Maybe smartphone is not the best name either, but at least it's better in my opinion.