Write files to the sass cache atomicly.
In multithreaded environments, the sass cache can be read in one
thread before the file is finished writing in another thread. This patch
causes the contents of a sass cache entry to be written to a tempfile
and then moved to the sass cache location when complete.

If several threads are reading the same sass cache entry simultanously
the will all miss the cache hit and the last writer will ultimately win,
meaning a small performance hit for this case but at least there will
not be any errors or data corruption.