Clean up the Sass function documentation.
This changes all references to parameters to be $-prefixed, adds
parameter and return type documentation for all parameters, and adds
parameter names to most assert_type calls. It also normalizes the
formatting and phrasing across all docstrings.

Hopefully now that this is more consistently formatted, it'll be
easier to follow the formatting when creating new functions.