Be a little less aggressive about redundant selector elimination.
Prior to this, when determining if a generated selector G that was
redundant with another selector S should be eliminated, the
specificity of S was compared to the minimum specificity of all
selectors that contributed to generating G. If S was as or more
specific, G was eliminated. However, this meant that if one of the
selectors that contributed to the generation of G had a higher
specificity than S, G could still be eliminated. This algorithm has
been adjusted to compare against the maximum specificity of the
selectors that contributed to G instead.

In addition, the specificity of a placeholder selector has been
increased to match that of a class selector.