Compute column locations based on an array of newline positions
Using `rindex` works well when we can guarantee that a newline will
occur reasonably close to the end of the string; this does not hold
true for source files that have already been minimized.

When processing source that has few newlines, we have to walk a very
long way through the entire string. This is compounded by the fact
that we may have to walk the same string many times, once for each
subsequent directive we are processing. This leads to behavior that
looks O(n^2).

Instead, we lazily pay a one-time cost to figure out where all the
newlines are, then quickly scan through those locations when
calculating a specific column.

This will be most significant on JRuby, which currently (1.7.5) has a
less-than-efficient implementation of `rindex` when non-ASCII
characters are involved. This fix speeds up our real-life asset
precompilation from 4:30 to 2:00!

Some synthetic benchmarks that worst expose the problem:

MRI (1.9.3-p448) - 250KiB of CSS
Before: 71.5
After:  24.8 (34.6%)

JRuby (1.7.5) - 30KiB of CSS
Before: 188.4
After:   27.3 (14.5%)