Move the sourcemap filename warning.
The location of the sourcemap warning was tightly coupled to the calling
code. This made it impossible for an importer to have source files for
which there is no public_url without a warning from Sass.

This patch updates Sass so that the importer can disable the warning by
simply returning nil from public_url without calling
`warn_about_public_url(uri)` which is now provided by the
`Importer::Base` class.