Support ids in SassScript.
To be completely correct in our implementation, we would need to
ensure that no color that is also a valid id changes its
representation, even in compressed mode. However, this would
negatively affect all users of Sass for the benefit of a narrow edge
case in an already low-usage section of CSS. As such, I've decided not
to preserve the representation of these colors in compressed mode.

Closes #1001