Do not freeze `true` or `false` in `Script::Value::Base.new`.
This fixes #1210, wherein the freezing of `true` causes strange side
effects in other gems. It also separates the `value.freeze` call from
the assignment `@value = value`, to emphasize that `value.freeze` is a
side effect that mutates the value, not a function that returns a frozen
copy of the value.