Store line number in Sequence#line=
The `scss-lint` project performs a number of static checks against SCSS
code to find style offenses.

`Sequence` objects were returning `nil` when calling `#line`. This was
because `Sequence#line=` was not recording the actual line in an
instance variable like `AbstractSequence`. Fix this by ensuring the line
number is stored in `@line`.

This will allow the `scss-lint` project to implement
https://github.com/brigade/scss-lint/issues/456