Add compatibility with listen 3.
It looks like it doesn't have a substantially different API than listen
2 in terms of what we use, so we just need to increase our version
constraint.

This also fixes the issue that was causing the error in #1766. We
required listen/version before checking whether we should add
vendor/listen/lib to the path, so we could end up looking at a version
constant that didn't represent the version we were actually using.

Closes #1766