First class functions.
First class functions are Sass values that refer to the current function
definition within its current scope. Even if that function is
overwritten it is still possible to call the older version by capturing
the function reference as an intermediate step.

The `call()` function now accepts function references and passing
function names as strings is deprecated. This prepares the way for the
module system which allows different modules to have functions of the
same name; without first class functions it wouldn't be possible for
a call site in one module to invoke a function defined in another via
dynamic lookup.

To get a reference to a function, use the `get-function($name)`
function. It will return a function reference or raise an error if no
sass function is defined. It is possible to create references to css
functions by passing `$css: true` as a second argument to
`get-function()`. CSS functions when called, will format their arguments
as strings passed to a CSS function call. Passing `$css: true` always
returns a CSS function even if there is already a Sass function of that
name.

Developers of frameworks that use `call` should not do the function
lookup for callers of their framework; this is likely to result in a
situation where the framework cannot resolve the function in 4.0.