Add multiline expressions
Enables splitting a long definition into several lines:

```sass
.shadow
  box-shadow: \
    5px 5px 10px black,
    inset 0 0 10px white

```

Symbol `\` is replaced with a concatenation of lines in a nested block.
This can be usefull for long css properties, variable maps or variable
lists. `\` has to be at the end of a line to work like this.

```sass
$rgb-colors: \
  red
  green
  blue
```

Transformation is done before running a parser, during building of a
lines array. It works in a similar fashion to processing block comments.

This pull request is related to issue #216