Correct error in docs for `mix`.
For better or worse the `$weight` argument is "how much of `$color1` to include
in the returned color" and not "how much to transition from `$color1` to
`$color2`". The rest of the docs get this right.

Here's a simple example that will show you that 100% favors `$color1`
and 0% favors `$color2`:

  body {
    background-color: mix(white, black, 0%); // black
    color: mix(white, black, 100%); // white
  }

Which outputs:

  body {
    background-color: black;
    color: white;
  }

See it in action here:

  http://www.sassmeister.com/gist/b805bc9bbc6e1f64c4a70899b0c9d9e8