Document asymmetric error handling. (#33)
The spec originally required that both the host and compiler send each
other ProtocolErrors whenever any errors are detected. The spec should
instead be such that the compiler reports ProtocolErrors, but the host
does not report to the compiler. Instead, the host should surface
errors to its consumers, since the host will most likely be used as a
library.