Make LegacyAsyncFunction type more usable
TypeScript requires that a callback type with a rest parameter be
passed a callback with a rest parameter. This makes it awkward to pass
in functions with a known number of parameters, as in
`foo($arg): (value, done) => { /* ... */ }`. This type declaration
makes that easier (although callers still need to explicitly declare a
type for the `done` parameter).