[Color 4] Explicitly define associated units for color channels (#3696)
"The associated unit" was referenced several times in the spec, most
notably for when to add a unit for `color.channel()`, but wasn't
explicitly defined anywhere. This is potentially ambiguous because any
linear channel can be defined using a percent or a unitless number,
but different ones are conventionally referred to as either one or the
other.