[Color 4] Treat missing channels as distinct from 0 for `==`
Now that `color.same()` explicitly treats missing channels as 0, it
makes more sense for `==` to fill the stricter role and treat them as
different. This is especially true because it already
requires (non-legacy) colors to be in the same space, so missing
channels are much more consistently meaningful than they would be in
cross-space comparisons.