Further improve documentation around importers in the entrypoint
This also updates the spec to match the changes in
sass/dart-sass#2208. This behavior is already largely the actual
behavior of both the JS and Dart APIs, so it doesn't need a full
proposal process. The only difference is that Dart Sass (but not the
JS API) currently accepts relative `url` arguments to
`compileString()`, but because the URL was already as a canonical URL,
we consider it implicitly expected to be absolute anyway.